+++ src/main/java/be/kuleuven/cs/jli40d/core/model/PlayerHand.java
@@ -17,18 +17,18 @@ public class PlayerHand implements Serializable
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
 
     @OneToMany
     @MapKey( name = "username" )
     private List<Card> playerHands;
 
-    public long getId()
+    public int getId()
     {
         return id;
     }
 
-    public void setId( long id )
+    public void setId( int id )
     {
         this.id = id;
     }
